+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/models/operations/get_integration_branding.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 import dataclasses
 import requests as requests_http
 from ..shared import branding as shared_branding
+from ..shared import errormessage as shared_errormessage
 from typing import Optional
 
 
@@ -23,6 +24,8 @@ class GetIntegrationBrandingResponse:
     r"""HTTP response status code for this operation"""
     branding: Optional[shared_branding.Branding] = dataclasses.field(default=None)
     r"""Success"""
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""