+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/models/operations/get_sync_status.py
@@ -3,6 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
+from ..shared import errormessage as shared_errormessage
+from ..shared import syncsummary as shared_syncsummary
 from typing import Optional
 
 
@@ -20,7 +22,11 @@ class GetSyncStatusResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
+    sync_summary: Optional[shared_syncsummary.SyncSummary] = dataclasses.field(default=None)
+    r"""Success"""