+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/models/operations/list_companies.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 import dataclasses
 import requests as requests_http
 from ..shared import companies as shared_companies
+from ..shared import errormessage as shared_errormessage
 from typing import Optional
 
 
@@ -29,6 +30,8 @@ class ListCompaniesResponse:
     r"""HTTP response status code for this operation"""
     companies: Optional[shared_companies.Companies] = dataclasses.field(default=None)
     r"""Success"""
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""Your `query` parameter was not correctly formed"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""