+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.2.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [PyPI v0.2.0] https://pypi.org/project/codat-sync-for-commerce-version-1/0.2.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-11-06 01:25:56
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.3.0] previous-versions/sync-for-commerce-version-1
+### Releases
+- [PyPI v0.3.0] https://pypi.org/project/codat-sync-for-commerce-version-1/0.3.0 - previous-versions/sync-for-commerce-version-1